Use common error when removing an unwatched file (#460)
 #### What does this pull request do?
The errors returned by the various implementations of the watcher are all different
which makes handling them difficult. This PR follows the suggestion in:
https://github.com/fsnotify/fsnotify/pull/455#issuecomment-1146037157 by @mattn
to create a common error which is wrapped by the implementations.

Replaces: https://github.com/fsnotify/fsnotify/pull/455

 #### Where should the reviewer start?
fsnotify.go err definition

 #### How should this be manually tested?
Remove an unwatched file and check the error returned.

Close #455