windows: protect access to isClosed with mutex (#454)
isClosed may be read and modified concurrently on Windows if the Watcher is closed at the same time as a new entry being added. This change protects isClosed to prevent race conditions.

The order of fields for the Windows Watcher have been rearranged to group fields protected by the mutex together to make it easier for reviewers to understand at a glance which fields are expected to be accessed through the mutex.